Updates readme.md with a note about charging with set limit. (#26)
Charging cycle can go much further than bclm set value. For an instance, setting 77 would charge battery when system is up and running to about 80%. When same system is shut down or put in sleep, charging can go up to 91% but mostly up to 89% with same setting of 77. This is tested on MacBook Air (2017) with macOS Monterey v. 12.6.2 (21G320). This is happens even if you set persist value.